Don't strip version restrictions that are stated in conditionals used to define flags.
Those restrictions are usually very reliable and have been chosen deliberately
to allow Cabal to guess the flag configuration to use in the given environment.
If those restrictions are gone, that mechanism will no longer work. As an
example, consider the following snippet from pandoc-1.9.4.2.cabal:

    if flag(blaze_html_0_5)
        build-depends: blaze-html >=0.5 && <0.6,
                       blaze-markup >=0.5.1 && <0.6
        exposed: True
        buildable: True
    else
        build-depends: blaze-html -any
        exposed: True
        buildable: True

If the version restrictions from the first case are removed, then Cabal will
not figure set 'blaze_html_0_5' automatically when compiled with blaze-html
version 0.5 or later, and then the required module 'blaze-markup' is not going
to be linked, causing the build to fail.