Compatibility with Cabal >= 3.14
readGenericPackageDescription now uses interpretSymbolicPath internally
[1] which allows for passing in a CWD directory explicitly rather than
just inheriting it from the running process.

We want to preserve the old behavior and behave the same as
writeGenericPackageDescription which does not use interpretSymbolicPath,
so we pass Nothing. This amounts to the same behavior as just returning
the underlying path (which is what we had before) [2] [3].

This also seems safer since readGenericPackageDescription seems to
assume that the passed in CWD directory matches the process working
directory [1].

[1]: https://github.com/haskell/cabal/blob/Cabal-v3.14.1.0/Cabal/src/Distribution/Simple/PackageDescription.hs#L83-L84
[2]: https://github.com/haskell/cabal/blob/Cabal-v3.14.1.0/Cabal-syntax/src/Distribution/Utils/Path.hs#L288-L305
[3]: https://github.com/haskell/cabal/blob/Cabal-v3.14.1.0/Cabal-syntax/src/Distribution/Utils/Path.hs#L257-L259