.github/workflows: create one job per GHC version
The runners don't have enough disk space to download all GHC versions we
are testing.

The solution is to run an initial job that writes a JSON list of all
attributes that ci.nix exposes to a GHA output. This can then be re-used
in a matrix to create one job per attribute of ci.nix. This means that
the targets built in parallel on independent runners which should reduce
CI time overall and solves the disk space problem. Only downside seems
to be that we need to wait on the Nix installer twice.